ANO file definition change
Ano file data lines changed to accommodate auxiliary information from FasTAN for tandem arrays.

gene_core is now the union of all the version used by EWM and Uncompress no longer needs byte alignment, simplifying sequence fetch.

A couple of small bugs fixed.

+++ GDBtoFA.c
@@ -192,7 +192,7 @@ int main(int argc, char *argv[])
   { GDB_CONTIG   *ctg;
     GDB_SCAFFOLD *scf;
     ANO_PAIR     *mask;
-    int64        *moff;
+    int          *moff;
     char         *hdr;
     int           nscf;
     char         *contig, *header;
